+++ app/alpha/courses/page.tsx
@@ -1,17 +1,11 @@
-// @TODO remove head and add appropriate metadata
+// @TODO add appropriate metadata
 
-import Head from "next/head";
 import CoursePreview from "@/components/Course/CoursePreview";
 
 /**
  * Feel free to update this interface when the Course model
  * has been clearly defined in the Prisma schema.
  */
-interface Course {
-  name: string;
-  description: string;
-  thumbnail: string;
-}
 
 const Courses = () => {
   const courses = [
@@ -26,10 +20,6 @@ const Courses = () => {
 
   return (
     <>
-      <Head>
-        <meta name="robots" content="noindex" />
-      </Head>
-
       <div className="mx-2">
         <div className="mt-8 flex max-w-5xl items-center justify-between border-b pb-4 sm:mx-auto sm:max-w-2xl lg:max-w-5xl">
           <h1 className="text-3xl font-extrabold tracking-tight text-neutral-50 sm:text-4xl ">
