+++ app/alpha/courses/web-development/page.tsx
@@ -1,6 +1,4 @@
-// @TODO remove head and add appropriate metadata
-
-import Head from "next/head";
+// @TODO add appropriate metadata
 
 import {
   CourseContens,
@@ -21,10 +19,6 @@ const contentsMock: IContent[] = [
 const WebDevelopmentCourse = () => {
   return (
     <>
-      <Head>
-        <meta name="robots" content="noindex" />
-      </Head>
-
       <div className="mx-2">
         <div className="mt-8 flex max-w-5xl items-center justify-between border-b pb-4 sm:mx-auto sm:max-w-2xl lg:max-w-5xl">
           <h1 className="text-3xl font-bold tracking-tight text-neutral-50 sm:text-4xl">
