+++ app/create/[[...paramsArr]]/page.tsx
@@ -2,7 +2,9 @@ import { redirect } from "next/navigation";
 import Content from "./_client";
 import { getServerAuthSession } from "@/server/auth";
 
-// @TODO - Maybe add Metadata for this page
+export const metadata = {
+  title: "New post - Codú",
+};
 
 export default async function Page() {
   const session = await getServerAuthSession();
