+++ app/layout.tsx
@@ -17,6 +17,23 @@ export const metadata = {
   icons: [{ rel: "icon", url: "/favicon.ico" }],
   publisher: "Codú",
   applicationName: "Codú",
+  keywords: [
+    "programming",
+    "frontend",
+    "community",
+    "learn",
+    "programmer",
+    "article",
+    "Python",
+    "JavaScript",
+    "AWS",
+    "HTML",
+    "CSS",
+    "Tailwind",
+    "React",
+    "Angular",
+    "backend",
+  ],
   metadataBase: new URL("https://www.codu.co"),
   openGraph: {
     images: "/images/og/home-og.png",
