+++ app/my-posts/page.tsx
@@ -2,7 +2,10 @@ import { redirect } from "next/navigation";
 import Content from "./_client";
 import { getServerAuthSession } from "@/server/auth";
 
-// @TODO - Maybe add Metadata for this page
+export const metadata = {
+  title: "Bookmarked posts",
+  description: "Find all your bookmarked articles here.",
+};
 
 export default async function Page() {
   const session = await getServerAuthSession();
