+++ app/notifications/page.tsx
@@ -2,7 +2,13 @@ import { redirect } from "next/navigation";
 import Content from "./_client";
 import { getServerAuthSession } from "@/server/auth";
 
-// @TODO - Maybe add Metadata for this page
+// @TODO - Add notifications count to title:
+// `Notfications ${count ? `(${count}) 🔔` : "| No new notifications"}`
+// Suggestion - Maybe add dynamic favicon showing notifications by default?
+export const metadata = {
+  title: "Notifications | No new notifications",
+  description: "You notification page. Look at your latest notifications.",
+};
 
 export default async function Page() {
   const session = await getServerAuthSession();
